requirements.txt
brython==3.5.1
Django==2.0.4
djangorestframework==3.8.2
pytz==2018.4
